新人大会用7人レイアウトの描画を追加、PlayerDropdown、Timer関連のコンポーネントを別ファイルに分割、ファイル名をcamelCaseに修正
