+++ src/browser/graphics/views/game-scene/16x9_3.tsx
@@ -1,5 +1,5 @@
 import { render } from '../../../render';
-import { ViewCanvas } from '../../components/view-canvas';
+import { ViewCanvas } from '../../components/viewCanvas';
 import { Footer } from '../../components/footer/index';
 import { Nameplate } from '../../components/nameplate';
 import { ViewSettings } from '../../../../types/viewsettings';
