+++ src/browser/graphics/views/game-scene/focus/5-player/16x9_5_5.tsx
@@ -1,11 +1,11 @@
-import { render } from '../../../../render';
-import { ViewCanvas } from '../../../components/view-canvas';
-import { Footer } from '../../../components/footer/index';
-import { Nameplate } from '../../../components/nameplate';
-import { ViewSettings } from '../../../../../types/viewsettings';
-import { MainTimer } from '../../../components/timer';
-import { Logo } from '../../../components/logo';
-import { Box } from '../../../components/box';
+import { render } from '../../../../../render';
+import { ViewCanvas } from '../../../../components/viewCanvas';
+import { Footer } from '../../../../components/footer/index';
+import { Nameplate } from '../../../../components/nameplate';
+import { ViewSettings } from '../../../../../../types/viewsettings';
+import { MainTimer } from '../../../../components/timer';
+import { Logo } from '../../../../components/logo';
+import { Box } from '../../../../components/box';
 
 const viewSettings: ViewSettings = {
   player: 5,
@@ -47,7 +47,7 @@ const App: React.FC = () => {
             borderRadius: 15,
           }}></Box>
         <Nameplate
-          index={3}
+          index={4}
           view={viewSettings}
           position={{ x: viewSettings.initPos.x, y: viewSettings.initPos.y + viewSettings.height }}
           mainView={{ width: 973, height: 80, x: 15, y: 578 }}
@@ -80,7 +80,7 @@ const App: React.FC = () => {
           customHeight={55}
           customFontSize={1.6}></Nameplate>
         <Nameplate
-          index={4}
+          index={3}
           view={viewSettings}
           position={{
             x: viewSettings.initPos.x + viewSettings.width + viewSettings.margin.x,
